refactor route functions from index to routeHandler
