+++ src/server/index.js
@@ -3,9 +3,7 @@ import bodyParser from 'body-parser'
 
 import params from '../../params'
 import routes from './constants/routes'
-import * as gameHandler from './eventHandlers/gameHandler'
-import * as idHandler from './eventHandlers/idHandler'
-import Player from './controllers/player'
+import * as routeHandler from './eventHandlers/routeHandler'
 
 const app = express()
 const server = require('http').createServer(app)
@@ -25,52 +23,28 @@ app.get('/', (req, res) => {
 })
 
 io.on('connection', (client) => {
-	client.on(routes.LOGIN, (userInfo) => {
-		let player = new Player(userInfo.id, client.id)
 
-		onlineUsers.push(player)
+    client.on(routes.LOGIN, (userInfo) => {
+        routeHandler.login(userInfo, client, onlineUsers)
 	})
+
     client.on(routes.CREATE_GAME, (gameName) => {
-        const gameId = idHandler.getGameId(gameName)
-        const id = gameHandler.findGame(gameId, activeGames)
-
-         if (id !== undefined) {
-            io.to(client.id).emit(routes.GAME_EXISTS, 'KO')
-        } else {
-            let game = gameHandler.createGame(client.id, onlineUsers)
-
-            game.setRoomInfo(gameId, gameName)
-            activeGames.push(game)
-            io.to(client.id).emit(routes.GAME_EXISTS, 'OK')
-        }
+        routeHandler.createGame(io, client, activeGames, onlineUsers, gameName)
     })
     client.on(routes.JOIN_GAME, (gameName) => {
-        const challenger = gameHandler.findPlayer(client.id, onlineUsers)
-        const gameId = idHandler.getGameId(gameName)
-        let game = gameHandler.findGame(gameId, activeGames)
-
-        game.setChallenger(challenger)
+        routeHandler.joinGame(client, onlineUsers, gameName, activeGames)
     })
+
     client.on(routes.START_GAME, () => {
-        let game = gameHandler.findGameBySocketId(client.id, activeGames)
-
-        /*if (game.waitingForPlayers()) {
-            io.to(client.id).emit('gameStarted', 'KO')
-            return
-        }*/
-        game.boardMaster = gameHandler.initBoard()
-        game.boardChallenger = gameHandler.initBoard()
-        io.to(game.master.socketID).emit('gameStarted', game.boardMaster)
-        //io.to(game.challenger.socketID).emit('gameStarted', game.boardChallenger)
+        routeHandler.startGame(io, client, activeGames)
     })
+
     client.on(routes.REQUEST_SHAPE, () => {
-        let game = gameHandler.findGameBySocketId(client.id, activeGames)
-        let shape = gameHandler.getShape(game, client.id)
-
-        io.to(client.id).emit(routes.EMITTED_SHAPE, shape)
+        routeHandler.requestShape(io, client, activeGames)
     })
+
     client.on('disconnect', () => {
-        console.log('user is disconnecting')
+        routeHandler.disconnect()
     })
 })
